[SDK-519] Refactor reveal api callback
